refactor: extract AuthorMeta and CommentSection components from ArticlePage
Unify divergent author-info patterns between ArticlePreview and ArticlePage
into a shared AuthorMeta component with variant prop (default/banner).
Extract comment form, list, and auth prompt into CommentSection component.
Disable NuGetAudit to unblock local dev builds with transitive vulnerability.
